+++ pypeit/core/arc.py
@@ -78,8 +78,33 @@ def setup_param(spectro_class, msarc_shape, fitstbl, arc_idx,
     # Return
     return arcparam
 
+def get_censpec(slit_left, slit_righ, slitpix, arcimg, inmask = None, box_rad = 3, xfrac = 0.5):
 
-def get_censpec(lordloc, rordloc, pixlocn, frame, det, nonlinear_counts=None, gen_satmask=False):
+    if inmask is None:
+        inmask = (slitpix > 0)
+
+    nslits = slit_left.shape[1]
+    (nspec, nspat) = arcimg.shape
+    maskslit = np.zeros(nslits, dtype=np.int)
+    trace = slit_left + xfrac*(slit_righ - slit_left)
+    arc_spec = np.zeros((nspec, nslits))
+
+    for islit in range(nslits):
+        msgs.info("Extracting an approximate arc spectrum at the centre of slit {:d}".format(islit + 1))
+        # Create a mask for the pixels that will contribue to the arc
+        spat_img = np.outer(np.ones(nspec), np.arange(nspat))  # spatial position everywhere along image
+        trace_img = np.outer(trace[:,islit], np.ones(nspat))  # left slit boundary replicated spatially
+        arcmask = (slitpix > 0) & inmask & (spat_img > (trace_img - box_rad)) & (spat_img < (trace_img + box_rad))
+        this_mean, this_med, this_sig = sigma_clipped_stats(arcimg, mask=~arcmask, sigma=3.0, axis=1)
+        arc_spec[:,islit] = this_med.data
+        if not np.any(arc_spec[:,islit]):
+            maskslit[islit] = 1
+
+    return arc_spec, maskslit
+
+
+# ToDO this code needs to be replaced. It is not masking outliers, and zeros out orders that leave the detector
+def get_censpec_old(lordloc, rordloc, pixlocn, frame, det, nonlinear_counts=None, gen_satmask=False):
     """ Extract a simple spectrum down the center of each slit
     Parameters
     ----------
@@ -442,7 +467,7 @@ def detect_lines(censpec, nfitpix=5, sigdetect = 10.0, FWHM = 10.0, cont_samp = 
 
     #         sigma finite  & sigma positive &  sigma < FWHM/2.35 & cen positive  &  cen on detector
     # TESTING
-    good = (~np.isnan(twid)) & (twid > 0.0) & (twid < FWHM/2.35) & (tcent > 0.0) & (tcent < xrng[-1]) & (tampl < nonlinear_counts)
+    good = (np.invert(np.isnan(twid))) & (twid > 0.0) & (twid < FWHM/2.35) & (tcent > 0.0) & (tcent < xrng[-1]) & (tampl < nonlinear_counts)
     ww = np.where(good)
     if debug:
         # Interpolate for bad lines since the fitting code often returns nan