+++ pypeit/flatfield.py
@@ -15,7 +15,10 @@ from pypeit import masterframe
 from pypeit.core import flat
 from pypeit import ginga
 from pypeit.par import pypeitpar
+from pypeit.core import pixels
 from pypeit.core import trace_slits
+from pypeit.core import tracewave
+
 
 from pypeit import debugger
 
@@ -63,7 +66,7 @@ class FlatField(processimages.ProcessImages, masterframe.MasterFrame):
     frametype = 'pixelflat'
 
     def __init__(self, spectrograph, file_list=[], det=1, par=None, setup=None, master_dir=None,
-                 mode=None, flatpar=None, msbias=None, msbpm=None, tslits_dict=None, tilts=None):
+                 mode=None, flatpar=None, msbias=None, msbpm = None, tslits_dict=None, tilts_dict=None):
 
         # Image processing parameters
         self.par = pypeitpar.FrameGroupPar(self.frametype) if par is None else par
@@ -74,13 +77,13 @@ class FlatField(processimages.ProcessImages, masterframe.MasterFrame):
 
         # MasterFrames: Specifically pass the ProcessImages-constructed
         # spectrograph even though it really only needs the string name
-        masterframe.MasterFrame.__init__(self, self.frametype, setup, master_dir=master_dir,
-                                         mode=mode)
+        masterframe.MasterFrame.__init__(self, self.frametype, setup,
+                                         master_dir=master_dir, mode=mode)
 
         # Parameters unique to this Object
         self.msbias = msbias
         self.tslits_dict = tslits_dict
-        self.tilts = tilts
+        self.tilts_dict = tilts_dict
         self.msbpm = msbpm
         if master_dir is None:
             self.master_dir = os.getcwd()
@@ -170,6 +173,7 @@ class FlatField(processimages.ProcessImages, masterframe.MasterFrame):
         """
         return masters.load_master_frame('illumflat', self.setup, self.mdir)
 
+    # ToDO this routine is deprecated and no longer called
     def slit_profile(self, slit):
         """
         Generate the slit profile for a given slit
@@ -201,7 +205,7 @@ class FlatField(processimages.ProcessImages, masterframe.MasterFrame):
         slordloc = self.tslits_dict['lcen'][:,slit]
         srordloc = self.tslits_dict['rcen'][:,slit]
         modvals, nrmvals, msblaze_slit, blazeext_slit, iextrap_slit \
-                = flat.slit_profile(slit, self.mspixelflat, self.tilts, slordloc, srordloc,
+                = flat.slit_profile(slit, self.mspixelflat, self.tilts_dict['tilts'], slordloc, srordloc,
                                       self.tslits_dict['slitpix'], self.tslits_dict['pixwid'],
                                       ntckx=self.ntckx, ntcky=self.ntcky)
         # Step
@@ -211,7 +215,7 @@ class FlatField(processimages.ProcessImages, masterframe.MasterFrame):
         # Return
         return modvals, nrmvals, msblaze_slit, blazeext_slit, iextrap_slit
 
-    def run(self, debug=False, show=False):
+    def run(self, debug = False, show = False):
         """
         Main driver to generate normalized flat field and illumination flats
 
@@ -246,44 +250,71 @@ class FlatField(processimages.ProcessImages, masterframe.MasterFrame):
         self.msillumflat = np.ones_like(self.rawflatimg)
         self.flat_model = np.zeros_like(self.rawflatimg)
 
+        final_tilts = np.zeros_like(self.rawflatimg)
+
         # Loop on slits
         for slit in range(self.nslits):
             msgs.info("Computing flat field image for slit: {:d}".format(slit + 1))
             thismask = (self.tslits_dict['slitpix'] == slit + 1)
             if self.msbpm is not None:
-                inmask = thismask & ~self.msbpm
+                inmask = ~self.msbpm
             else:
-                inmask = thismask
+                inmask = np.ones_like(self.rawflatimg,dtype=bool)
+
             # Fit flats for a single slit
-            self.mspixelflat[thismask], self.msillumflat[thismask], self.flat_model[thismask] \
-                    = flat.fit_flat(self.rawflatimg, self.tilts, thismask,
-                                    self.tslits_dict['lcen'][:,slit],
-                                    self.tslits_dict['rcen'][:,slit], inmask=inmask, debug=debug)
+            this_tilts_dict = {'tilts':self.tilts_dict['tilts'], 'coeffs':self.tilts_dict['coeffs'][:,:,slit],
+                               'func2D':self.tilts_dict['func2D']}
+            nonlinear_counts = self.spectrograph.detector[self.det - 1]['nonlinear']*\
+                               self.spectrograph.detector[self.det - 1]['saturation']
+            pixelflat, illumflat, flat_model, thismask_out, slit_left_out, slit_righ_out = \
+                flat.fit_flat(self.rawflatimg, this_tilts_dict, thismask,self.tslits_dict['lcen'][:, slit],
+                              self.tslits_dict['rcen'][:,slit],inmask=inmask,tweak_slits = self.flatpar['tweak_slits'],
+                              nonlinear_counts=nonlinear_counts, debug=debug)
+            self.mspixelflat[thismask_out] = pixelflat[thismask_out]
+            self.msillumflat[thismask_out] = illumflat[thismask_out]
+            self.flat_model[thismask_out] = flat_model[thismask_out]
+            # Did we tweak slit boundaries? If so, update the tslits_dict and the tilts_dict
+            if self.flatpar['tweak_slits']:
+                self.tslits_dict['lcen'][:, slit] = slit_left_out
+                self.tslits_dict['rcen'][:, slit] = slit_righ_out
+                this_tilts = tracewave.coeff2tilts(this_tilts_dict['coeffs'], self.rawflatimg.shape, self.tilts_dict['func2D'])
+                final_tilts[thismask_out] = this_tilts[thismask_out]
+
+        # If we tweaked the slits update the tslits_dict and the tilts_dict
+        if self.flatpar['tweak_slits']:
+            self.tilts_dict['tilts'] = final_tilts
+            # Update the tslits_dict
+            self.tslits_dict['slitpix']= pixels.slit_pixels(self.tslits_dict['lcen'], self.tslits_dict['rcen'], self.rawflatimg.shape, 0)
+            # ToDo no need to store the ximg and edgmask in the tslits_dict, they can be generated on the fly
+            ximg, edge_mask = pixels.ximg_and_edgemask(self.tslits_dict['lcen'], self.tslits_dict['rcen'], self.tslits_dict['slitpix'])
+            self.tslits_dict['ximg'] = ximg
+            self.tslits_dict['edge_mask'] = edge_mask
 
         if show:
-            # Global skysub is the first step in a new extraction so
-            # clear the channels here
+            # Global skysub is the first step in a new extraction so clear the channels here
             self.show(slits=True, wcs_match = True)
 
         # Return
         return self.mspixelflat, self.msillumflat
 
+
+
     def show(self, slits = True, wcs_match = True):
-        viewer, ch = ginga.show_image(self.mspixelflat, chname='pixeflat', cuts=(0.9, 1.1),
-                                      wcs_match=wcs_match, clear=True)
-        viewer, ch = ginga.show_image(self.msillumflat, chname='illumflat', cuts=(0.9, 1.1),
-                                      wcs_match=wcs_match)
+
+        viewer, ch = ginga.show_image(self.mspixelflat, chname='pixeflat', cuts=(0.9, 1.1), wcs_match=wcs_match, clear=True)
+        viewer, ch = ginga.show_image(self.msillumflat, chname='illumflat', cuts=(0.9, 1.1), wcs_match=wcs_match)
         viewer, ch = ginga.show_image(self.rawflatimg, chname='flat', wcs_match=wcs_match)
         viewer, ch = ginga.show_image(self.flat_model, chname='flat_model', wcs_match=wcs_match)
 
+
         if slits:
             if self.tslits_dict is not None:
-                slit_ids = [trace_slits.get_slitid(self.rawflatimg.shape, self.tslits_dict['lcen'],
-                                                   self.tslits_dict['rcen'], ii)[0]
-                                for ii in range(self.tslits_dict['lcen'].shape[1])]
-                ginga.show_slits(viewer, ch, self.tslits_dict['lcen'], self.tslits_dict['rcen'],
-                                 slit_ids)
+                slit_ids = [trace_slits.get_slitid(self.rawflatimg.shape, self.tslits_dict['lcen'], self.tslits_dict['rcen'], ii)[0]
+                            for ii in range(self.tslits_dict['lcen'].shape[1])]
+                ginga.show_slits(viewer, ch, self.tslits_dict['lcen'], self.tslits_dict['rcen'], slit_ids)
 
+
+'''
     def show_old(self, attr='norm', display='ginga'):
         """
         Show one of the internal images
@@ -306,6 +337,9 @@ class FlatField(processimages.ProcessImages, masterframe.MasterFrame):
             if self.mspixelflatnrm is not None:
                 ginga.show_image(self.mspixelflatnrm)
 
+'''
+
+'''
     def run_old(self, armed=False):
         """
         Main driver to generate normalized flat field
@@ -383,22 +417,27 @@ class FlatField(processimages.ProcessImages, masterframe.MasterFrame):
         inslit = self.tslits_dict['slitpix'] >= 1.
         self.mspixelflatnrm[~inslit] = 1.
 
-#        # QA
-#        if np.array_equal(self._idx_flat, self._idx_trace):
-#            # The flat field frame is also being used to trace the slit
-#            # edges and determine the slit profile. Avoid recalculating
-#            # the slit profile and blaze function and save them here.
-#            self.SetFrame(self._msblaze, msblaze, det)
-#            self.SetFrame(self._slitprof, slit_profiles, det)
-#            masters.save_masters(self, det, mftype='slitprof')
-#            if settings.argflag["reduce"]["slitprofile"]["perform"]:
-#                msgs.info("Preparing QA of each slit profile")
-#                arproc.slit_profile_qa(self, mstracenrm, slit_profiles,
-#                                       self._lordloc[det - 1], self._rordloc[det - 1],
-#                                       self._slitpix[det - 1], desc="Slit profile")
-#            msgs.info("Saving blaze function QA")
-#            artracewave.plot_orderfits(self.setup, msblaze, flat_ext1d, desc="Blaze function")
+        # QA
+"""
+        if np.array_equal(self._idx_flat, self._idx_trace):
+            # The flat field frame is also being used to trace the slit edges and determine the slit
+            # profile. Avoid recalculating the slit profile and blaze function and save them here.
+            self.SetFrame(self._msblaze, msblaze, det)
+            self.SetFrame(self._slitprof, slit_profiles, det)
+            masters.save_masters(self, det, mftype='slitprof')
+            if settings.argflag["reduce"]["slitprofile"]["perform"]:
+                msgs.info("Preparing QA of each slit profile")
+                #                            arqa.slit_profile(self, mstracenrm, slit_profiles, self._lordloc[det - 1], self._rordloc[det - 1],
+                #                                              self._slitpix[det - 1], desc="Slit profile")
+                arproc.slit_profile_qa(self, mstracenrm, slit_profiles,
+                                       self._lordloc[det - 1], self._rordloc[det - 1],
+                                       self._slitpix[det - 1], desc="Slit profile")
+            msgs.info("Saving blaze function QA")
+            #                        arqa.plot_orderfits(self, msblaze, flat_ext1d, desc="Blaze function")
+            artracewave.plot_orderfits(self.setup, msblaze, flat_ext1d, desc="Blaze function")
+        #
+   """
 
         # Return
         return self.mspixelflatnrm, self.slit_profiles
-
+'''