+++ pypeit/pypeit.py
@@ -630,7 +630,7 @@ class MultiSlit(PypeIt):
                                                    maskslits=maskslits)
 
         # Global sky subtraction, first pass. Uses skymask from object finding
-        global_sky0 = self.sciI.global_skysub(self.caliBrate.tslits_dict, self.caliBrate.mstilts,
+        global_sky0 = self.sciI.global_skysub(self.caliBrate.tslits_dict, self.caliBrate.tilts_dict['tilts'],
                                               use_skymask=True, maskslits=maskslits, show=self.show)
 
         # Object finding, second pass on frame *with* sky subtraction. Show here if requested
@@ -641,7 +641,7 @@ class MultiSlit(PypeIt):
         vel_corr = None
         if nobj > 0:
             # Global sky subtraction second pass. Uses skymask from object finding
-            global_sky = self.sciI.global_skysub(self.caliBrate.tslits_dict, self.caliBrate.mstilts,
+            global_sky = self.sciI.global_skysub(self.caliBrate.tslits_dict, self.caliBrate.tilts_dict['tilts'],
                                                  use_skymask=True, maskslits=maskslits, show=self.show)
 
             skymodel, objmodel, ivarmodel, outmask, sobjs = self.sciI.local_skysub_extract(self.caliBrate.mswave, maskslits=maskslits,